Criando o bottom sheet de aviso

+++ app/src/main/java/com/devspace/taskbeats/MainActivity.kt
@@ -52,11 +52,24 @@ class MainActivity : AppCompatActivity() {
         }
 
         categoryAdapter.setOnLongClickListener { categoryToBeDeleted ->
-            val categoryEntityToBeDeleted = CategoryEntity(
-                    name = categoryToBeDeleted.name,
-                    isSelected = categoryToBeDeleted.isSelected
-            )
-            deleteCategory(categoryEntityToBeDeleted)
+            if (categoryToBeDeleted.name != "+"){
+                val title: String = this.getString(R.string.category_delete_title)
+                val description: String = this.getString(R.string.category_delete_description)
+                val btnText: String = this.getString(R.string.delete)
+
+                showinfobottomsheet(
+                    title,
+                    description,
+                    btnText
+                ){
+                    val categoryEntityToBeDeleted = CategoryEntity(
+                        name = categoryToBeDeleted.name,
+                        isSelected = categoryToBeDeleted.isSelected
+                    )
+                    deleteCategory(categoryEntityToBeDeleted)
+                }
+            }
+
         }
 
         categoryAdapter.setOnClickListener { selected ->
@@ -104,6 +117,25 @@ class MainActivity : AppCompatActivity() {
         GlobalScope.launch(Dispatchers.Main) {
             getTasksFromDataBase()
         }
+    }
+
+    private fun showinfobottomsheet(
+        title: String,
+        description: String,
+        btnText: String,
+        onClick: () -> Unit
+    ) {
+        val infoBottomSheet = InfoBottomSheet(
+            title = title,
+            description = description,
+            btnText = btnText,
+            onClick
+        )
+
+        infoBottomSheet.show(
+            supportFragmentManager,
+                "infoBottomSheet")
+
     }
 
     private fun getCategoriesFromDataBase() {
